updates: add update to next stable option
This is to remove some of the issues around auto updating to stable from the
nightly builds. It also give the user more control over auto updates since
the user can now specify a preference whether or not to auto return to stable
builds when the next stable build is available.